Remove previous matching-service and replace with new matching service.
