--- services/matching-service/entry-points/api/server.ts
@@ -1,47 +1,0 @@
-import express from "express";
-import type { Server } from "http";
-import type { AddressInfo } from "net";
-
-// import { errorHandler } from "../../errorHandler";
-
-import defineRoutes from "./routes";
-
-const cors = require("cors");
-
-let connection: Server;
-
-// ️️️✅ Best Practice: API exposes a start/stop function to allow testing control WHEN this should happen
-async function openConnection(
-  expressApp: express.Application,
-): Promise<AddressInfo> {
-  return new Promise((resolve) => {
-    const webServerPort = process.env.SERVER_PORT || 4001;
-    connection = expressApp.listen(webServerPort, () => {
-      resolve(connection.address() as AddressInfo);
-    });
-  });
-}
-
-async function startWebServer(): Promise<AddressInfo> {
-  // ️️️✅ Best Practice: Declare a strict configuration schema and fail fast if the configuration is invalid
-  const expressApp = express();
-  expressApp.use(cors());
-  expressApp.use(express.urlencoded({ extended: true }));
-  expressApp.use(express.json());
-  defineRoutes(expressApp);
-//   expressApp.use(errorHandler);
-  const APIAddress = await openConnection(expressApp);
-  return APIAddress;
-}
-
-async function stopWebServer() {
-  return new Promise<void>((resolve) => {
-    if (connection !== undefined) {
-      connection.close(() => {
-        resolve();
-      });
-    }
-  });
-}
-
-export { startWebServer, stopWebServer };
