+++ services/matching-service/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "matching-service",
+  "name": "matching-socketio",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "dev": "ts-node-dev start.ts",
@@ -34,6 +34,7 @@
     "@types/express": "^4.17.17",
     "cors": "^2.8.5",
     "express": "^4.18.2",
+    "socket.io": "^4.7.2",
     "zod": "^3.22.2"
   }
 }
