--- services/matching-service/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Node 20",
-  "_version": "20.1.0",
-
-  "compilerOptions": {
-    "strictNullChecks": true,
-    "lib": ["es2023"],
-    "module": "node16",
-    "target": "es2022",
-
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node16"
-  }
-}
