--- services/matching-socketio/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "matching-socketio",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "ts-node-dev start.ts",
-    "lint:fix": "eslint --ext .ts --fix ."
-  },
-  "devDependencies": {
-    "@prisma/client": "^5.4.1",
-    "@types/node": "^20.8.3",
-    "@typescript-eslint/eslint-plugin": "^6.7.4",
-    "@typescript-eslint/parser": "^6.7.4",
-    "bun-types": "latest",
-    "eslint": "^8.51.0",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-config-airbnb-typescript": "^17.1.0",
-    "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-import": "^2.28.1",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-unused-imports": "^3.0.0",
-    "pino": "^8.15.6",
-    "prettier": "^3.0.3",
-    "prisma": "^5.4.1",
-    "ts-node": "^10.9.1",
-    "ts-node-dev": "^2.0.0",
-    "tslib": "^2.6.2",
-    "typescript": "^5.2.2"
-  },
-  "peerDependencies": {
-    "typescript": "^5.0.0"
-  },
-  "dependencies": {
-    "@types/express": "^4.17.17",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "socket.io": "^4.7.2",
-    "zod": "^3.22.2"
-  }
-}
